config: T2485: move configsession.py to conf/session.py
