+++ python/vyos/config/session.py
