+++ src/conf_mode/host_name.py
@@ -167,9 +167,9 @@ def apply(config):
     # restart pdns if it is used - we check for the control dir to not raise
     # an exception on system startup
     #
-    #   File "/usr/lib/python3/dist-packages/vyos/configsession.py", line 128, in __run_command
+    #   File "/usr/lib/python3/dist-packages/vyos/config/session.py", line 128, in __run_command
     #     raise ConfigSessionError(output)
-    # vyos.configsession.ConfigSessionError: [ system domain-name vyos.io ]
+    # vyos.config.session.ConfigSessionError: [ system domain-name vyos.io ]
     # Fatal: Unable to generate local temporary file in directory '/run/powerdns': No such file or directory
     if os.path.isdir('/run/powerdns'):
         ret = run('/usr/bin/rec_control --socket-dir=/run/powerdns ping')
